+++ src/transformer.js
@@ -67,13 +67,13 @@ class Transformer {
    */
   static dumpObject (obj, name = 'noname') {
     if (V.isUndefined(obj)) {
-      return `${name}: undefined`
+      return `${name}:undefined`
     } else if (V.isNull(obj)) {
-      return `${name}: null`
+      return `${name}:null`
     } else if (V.isObject(obj) || V.isKeyedCollection(obj)) {
-      return `${name}: ${JSON.stringify(obj)}`
+      return `${name}:${JSON.stringify(obj)}`
     } else {
-      return `${name}: '${obj.toString()}'`
+      return `${name}:'${obj.toString()}'`
     }
   }