+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Summary Changelog:
   and even with a boolean value, both with a non default datacontenttype (for example 'text/plain')
 - Feature: handle 'data' as array (it was previously forbidden)
 - Feature: update example and update/simplify README
+- Feature: update JSONBatch strict validation to consider even null/undefined items in the array
 
 ## [0.7.0](https://github.com/smartiniOnGitHub/cloudevent.js/releases/tag/0.7.0) (2020-09-20)
 Summary Changelog: